Moved the file access workaround to the EntityHelperTrait.
