+++ src/EntityEmbedDisplay/EntityReferenceFieldFormatter.php
@@ -9,7 +9,6 @@ namespace Drupal\entity_embed\EntityEmbedDisplay;
 
 use Drupal\Core\Field\FieldDefinition;
 use Drupal\Core\Session\AccountInterface;
-use Drupal\entity_embed\EntityHelperTrait;
 
 /**
  * Embed entity displays for entity_reference field formatters.
@@ -23,7 +22,6 @@ use Drupal\entity_embed\EntityHelperTrait;
  * )
  */
 class EntityReferenceFieldFormatter extends FieldFormatterEntityEmbedDisplayBase {
-  use EntityHelperTrait;
 
   /**
    * {@inheritdoc}
@@ -53,9 +51,6 @@ class EntityReferenceFieldFormatter extends FieldFormatterEntityEmbedDisplayBase
       case 'entity_reference_entity_view':
         // Cannot render an entity if it does not have a view controller.
         // @todo Remove when https://drupal.org/node/2204325 is fixed in core.
-        // The entity manager is not injected via dependency here because
-        // ideally we resolve the core issue here rather than a concern about
-        // proper testability here on a one-off basis.
         return $this->canRenderEntity($this->getContextValue('entity'));
 
       default:
