+++ src/EntityEmbedDisplay/FileFieldFormatter.php
@@ -8,7 +8,6 @@
 namespace Drupal\entity_embed\EntityEmbedDisplay;
 
 use Drupal\Core\Field\FieldDefinition;
-use Drupal\Core\Session\AccountInterface;
 
 /**
  * Embed entity displays for file field formatters.
@@ -39,42 +38,6 @@ class FileFieldFormatter extends EntityReferenceFieldFormatter {
     $value = parent::getFieldValue($definition);
     $value += array_intersect_key($this->getAttributeValues(), array('description' => ''));
     return $value;
-  }
-
-  /**
-   * {@inheritdoc}
-   */
-  public function access(AccountInterface $account = NULL) {
-    if (!$this->isValidEntityType()) {
-      return FALSE;
-    }
-
-    // Due to issues with access checking with file entities in core, we cannot
-    // actually use Entity::access() which would have been called by
-    // parent::access().
-    // @see https://drupal.org/node/2128791
-    // @see https://drupal.org/node/2148353
-    // @see https://drupal.org/node/2078473
-    switch (file_uri_scheme($this->getContextValue('entity')->getFileUri())) {
-      case 'public':
-        return TRUE;
-
-      case 'private':
-      case 'temporary':
-        $headers = $this->moduleHandler()->invokeAll('file_download', array($uri));
-        foreach ($headers as $result) {
-          if ($result == -1) {
-            return FALSE;
-          }
-        }
-
-        if (count($headers)) {
-          return TRUE;
-        }
-        break;
-    }
-
-    return FALSE;
   }
 
   /**
